Fixes remote indexing calls
There's a question around whether we should pass through the repo in the
case of the backend reporting the Repo available. At present the error
is passed through to the caller, which doesn't allow a graceful refill
of a front-end instance. That can be added as an explicit error handler
during indexRemote(), or the error(s) could be handled more gracefully
in general.