Add output context options for searches
Add the grep like -A, -B and -C options to `afind search` as well as the
server side functionality to implement it. New unit tests pass but wider
testing is required as although the zero context output has not changed,
-C=1 produces at least two lines of context, and nearby match lines have
an impact on the total number of context lines, suggesting some bleeding
between groups is occuring.