afind client: Use AFIND_SERVER from environment
Use the AFIND_SERVER environment variable if the --server argument is
not provided to the `afind` command line program. If the port number is
not provided after the address (e.g., 'afind' instead of 'afind:30800')
the default RPC port number 30800 will be used.

This change in handling of the server address value also applies to the
--server flag, which is a new (backwards compatible) change.